crewai-process-type/requirements.txt
crewai
crewai-tools
langfuse
python-dotenv
openlit